started updating "n_body" directory to be more modular an support opencl

+++ include/graphics/instanced_data.h
@@ -8,8 +8,8 @@
 namespace raw::graphics {
 struct instanced_data {
 	glm::mat4 model;
-	uint32_t  texture1;
-	uint32_t  texture2;
+	uint64_t  texture1;
+	uint64_t  texture2;
 };
 } // namespace raw::graphics
 #endif // SPACE_EXPLORER_INSTANCED_DATA_H
